removed api/queries & inference_gateway/queries folder (moved queries to new root folder) + fixed endpoint in ridges.py
