+++ api/src/endpoints/evaluation_sets.py
@@ -1,7 +1,7 @@
 from typing import List
 from fastapi import APIRouter
 from models.evaluation_set import EvaluationSetProblem
-from api.queries.evaluation_set import get_all_evaluation_set_problems_in_latest_set_id
+from queries.evaluation_set import get_all_evaluation_set_problems_in_latest_set_id
