+++ api/src/endpoints/validator.py
@@ -14,9 +14,9 @@ from pydantic import BaseModel
 from functools import wraps
 import api.config as config
 import utils.logger as logger
-from api.queries.agent import get_top_agents, get_agent_by_id, update_agent_status, get_next_agent_id_awaiting_evaluation_for_validator_hotkey
-from api.queries.evaluation import get_evaluation_by_id, get_hydrated_evaluation_by_id, update_evaluation_finished_at, create_new_evaluation_and_evaluation_runs, get_num_successful_validator_evaluations_for_agent_id, update_unfinished_evaluation_runs_in_evaluation_id_to_errored
-from api.queries.evaluation_run import get_evaluation_run_by_id, update_evaluation_run_by_id, \
+from queries.agent import get_top_agents, get_agent_by_id, update_agent_status, get_next_agent_id_awaiting_evaluation_for_validator_hotkey
+from queries.evaluation import get_evaluation_by_id, get_hydrated_evaluation_by_id, update_evaluation_finished_at, create_new_evaluation_and_evaluation_runs, get_num_successful_validator_evaluations_for_agent_id, update_unfinished_evaluation_runs_in_evaluation_id_to_errored
+from queries.evaluation_run import get_evaluation_run_by_id, update_evaluation_run_by_id, \
     get_all_evaluation_runs_in_evaluation_id, create_evaluation_run_log, check_if_evaluation_run_logs_exist
 from models.agent import Agent, AgentStatus
 from models.evaluation import Evaluation, EvaluationStatus
