+++ api/src/main.py
@@ -10,7 +10,7 @@ import uvicorn
 from api.loops.validator_heartbeat_timeout import validator_heartbeat_timeout_loop
 
 
-from api.queries.evaluation import set_all_unfinished_evaluation_runs_to_errored
+from queries.evaluation import set_all_unfinished_evaluation_runs_to_errored
 from api.src.endpoints.upload import router as upload_router
 from api.src.endpoints.retrieval import router as retrieval_router
 from api.src.endpoints.scoring import router as scoring_router
