+++ queries/inference.py
